+++ wcfsetup/install/files/lib/system/option/AboutMeOptionType.class.php
@@ -1,7 +1,6 @@
 <?php
 namespace wcf\system\option;
 use wcf\data\option\Option;
-use wcf\system\bbcode\PreParser;
 use wcf\system\exception\UserInputException;
 use wcf\system\message\censorship\Censorship;
 use wcf\system\WCF;
@@ -33,15 +32,5 @@ class AboutMeOptionType extends MessageOptionType {
 				throw new UserInputException($option->optionName, 'censoredWordsFound');
 			}
 		}
-	}
-	
-	/**
-	 * @inheritDoc
-	 */
-	public function getData(Option $option, $newValue) {
-		$newValue = parent::getData($option, $newValue);
-		
-		// run pre-parsing
-		return PreParser::getInstance()->parse($newValue);
 	}
 }